More strict table extraction
