+++ pysqldf/tests/__init__.py
@@ -1,2 +1,5 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
+
+from .sqldf import *
+from .load import *
